requirements.txt
aiohttp==3.9.5
fastapi==0.111.0
fastapi-cli==0.0.4
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
requests==2.32.2
uvicorn==0.29.0